Initial Public Exposure of bot
The internal history is not being exposed because I have hardcoded
passwords commited right through.

The bot is a touch of a hack.  It's aimed for a single use.

I started with Bot::BasicBot but started wanting to reimplement some
of the internals.  So I switched to POE::Component::IRC.

The Object structure I am currently using is an artifact of this
history.  I probably wouldn't choose if starting from scratch again.

I should also confess that I have never used POE before and don't really
understand it.